챌린지 상세조회, 좋아요, 스탬프, 댓글 등..

+++ src/main/java/com/sm/seoulmate/domain/attraction/dto/AttractionResponse.java
@@ -27,6 +27,6 @@ public record AttractionResponse(
         @Schema(description = "이미지 url")
         String imageUrl,
         @Schema(description = "좋아요 수")
-        Long likes
+        Integer likes
 ) {
 }
